+++ changelog.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Feature: Resolve Files and Layout entries in Resources Importer Template files
 - Feature: Inject SQL language into custom-sql definition files
 - Feature: OSGi component properties for Fragment components
+- Feature: Add support for custom fragment html tags and attributes
 - Bugfix: Exception when trying to inject Javascript inside aui:script-tags containing additonal attributes (like "use" or "require")
 - Bugfix: Avoid NullPointerException when trying to lookup documentation for an unknown OSGi property
 - Change: Fixed support for IntelliJ 2021.3